requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
asgiref==3.8.1
attrs==24.3.0
blinker==1.8.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
dj-rest-auth==7.0.0
Django==4.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
mysql-connector==2.2.9
mysql-connector-python==9.1.0
mysqlclient==2.2.4
packaging==24.2
pillow==11.0.0
propcache==0.2.1
psycopg2==2.9.10
PyJWT==2.9.0
PyMySQL==1.1.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
twilio==9.4.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.2
yarl==1.18.3
